+++ @udir-design/react/src/components/fileUpload/FileUploadItem.tsx
@@ -17,7 +17,6 @@ import {
 } from '@udir-design/icons';
 import { Button } from '../button';
 import { Card } from '../card';
-import { Divider } from '../divider';
 import { Link } from '../link';
 import { Spinner } from '../spinner';
 
@@ -124,104 +123,6 @@ export const FileUploadItem = forwardRef<HTMLDivElement, FileUploadItemProps>(
 );
 FileUploadItem.displayName = 'FileUpload.Item';
 
-export interface FileUploadListItem {
-  /**
-   * Either a native File or file metadata.
-   */
-  file: File;
-  /**
-   * href to file location.
-   */
-  href?: string;
-  /**
-   * Error message relating to the item.
-   */
-  error?: string;
-  /**
-   * Toggle loading state.
-   *
-   * @default false
-   */
-  loading?: boolean;
-  /**
-   * @default false
-   */
-  readonly?: boolean;
-}
-
-export interface FileUploadListProps extends Omit<
-  HTMLAttributes<HTMLDivElement>,
-  'data-color'
-> {
-  'data-size'?: Size;
-  /**
-   * List of file items to display.
-   */
-  items: FileUploadListItem[];
-  /**
-   * Callback when the remove button is clicked.
-   */
-  onRemove: (file: File, event: MouseEvent<HTMLButtonElement>) => void;
-}
-
-export const FileUploadList = forwardRef<HTMLDivElement, FileUploadListProps>(
-  ({ items, onRemove, className, 'data-size': size, ...rest }, ref) => (
-    <Card
-      ref={ref}
-      className={cl('uds-file-upload__list', className)}
-      data-size={size}
-      {...rest}
-    >
-      <ul>
-        {items.map((item, index) => (
-          <li
-            key={item.file.name}
-            className="uds-file-upload__list-item"
-            aria-busy={item.loading || undefined}
-          >
-            {index > 0 && <Divider />}
-            <div className="uds-file-upload__list-row">
-              <div className="uds-file-upload__list-icon">
-                <Icon
-                  file={item.file}
-                  showError={Boolean(item.error)}
-                  loading={item.loading}
-                />
-              </div>
-              <div className="uds-file-upload__list-content">
-                <FileName file={item.file} href={item.href} />
-                {!item.loading && (
-                  <span className="uds-file-upload__list-size">
-                    {formatFileSize(item.file)}
-                  </span>
-                )}
-                {item.error && (
-                  <span className="uds-file-upload__list-error">
-                    <XMarkOctagonFillIcon aria-hidden />
-                    {item.error}
-                  </span>
-                )}
-              </div>
-              {!item.loading && !item.readonly && (
-                <Tooltip content="">
-                  <Button
-                    icon
-                    onClick={(e) => onRemove(item.file, e)}
-                    variant="tertiary"
-                  >
-                    <TrashIcon aria-hidden />
-                  </Button>
-                </Tooltip>
-              )}
-            </div>
-          </li>
-        ))}
-      </ul>
-    </Card>
-  ),
-);
-FileUploadList.displayName = 'FileUpload.List';
-
 export function formatFileSize(file: File): string | null {
   if (!file.size) {
     return null;
@@ -231,7 +132,7 @@ export function formatFileSize(file: File): string | null {
   return `${megaBytes.toFixed(2)} MB`;
 }
 
-function Icon({
+export function Icon({
   file,
   showError,
   loading,
@@ -274,7 +175,7 @@ function Icon({
   }
 }
 
-const downloadFile = (file: File): void => {
+export const downloadFile = (file: File): void => {
   const a = document.createElement('a');
   const url = URL.createObjectURL(file);
   a.href = url;
@@ -289,7 +190,7 @@ interface FileNameProps {
   href?: string;
 }
 
-const FileName = ({ file, href }: FileNameProps) => {
+export const FileName = ({ file, href }: FileNameProps) => {
   if (href) {
     return <Link href={href}>{file.name}</Link>;
   }
