+++ @udir-design/react/src/components/fileUpload/index.ts
@@ -1,15 +1,13 @@
 import { FileUpload } from './FileUpload';
 import { FileUploadDropzone } from './FileUploadDropzone';
-import { FileUploadItem, FileUploadList } from './FileUploadItem';
+import { FileUploadItem } from './FileUploadItem';
+import { FileUploadList } from './FileUploadList';
 import { FileUploadTrigger } from './FileUploadTrigger';
 
 export type { FileUploadProps } from './FileUploadTrigger';
 export type { FileUploadDropzoneProps } from './FileUploadDropzone';
-export type {
-  FileUploadItemProps,
-  FileUploadListItem,
-  FileUploadListProps,
-} from './FileUploadItem';
+export type { FileUploadItemProps } from './FileUploadItem';
+export type { FileUploadListItem, FileUploadListProps } from './FileUploadList';
 export {
   FileUpload,
   FileUploadDropzone,
